build: 调整项目结构
