+++ shared/src/commonMain/kotlin/androidx/datastore/core/Closeable.kt
