+++ shared/src/wasmJsMain/kotlin/xyz/xfqlittlefan/fhraise/Platform.wasmJs.kt
@@ -16,27 +16,10 @@
  * with Fhraise. If not, see <https://www.gnu.org/licenses/>.
  */
 
-class JVMPlatform : Platform {
-    override val name: String = "Java ${System.getProperty("java.version")}"
+package xyz.xfqlittlefan.fhraise
+
+class WasmPlatform : Platform {
+    override val name: String = "Web with Kotlin/Wasm"
 }
 
-actual fun getPlatform(): Platform = JVMPlatform()
-
-enum class DesktopPlatform {
-    Linux, Windows, MacOS, Unknown;
-
-    companion object {
-        /**
-         * Identify OS on which the application is currently running.
-         */
-        val Current: DesktopPlatform by lazy {
-            val name = System.getProperty("os.name")
-            when {
-                name?.startsWith("Linux") == true -> Linux
-                name?.startsWith("Win") == true -> Windows
-                name == "Mac OS X" -> MacOS
-                else -> Unknown
-            }
-        }
-    }
-}
+actual fun getPlatform(): Platform = WasmPlatform()
